Cleaner exp scripts

+++ exp/exp_enhanced_sampling_alanine.py
@@ -1,3 +1,4 @@
+import os
 import numpy as np
 import sys
 sys.path.append('/home/dominic/PycharmProjects/CV_learning')
@@ -7,7 +8,7 @@ from EnhancedSamplingExperiments import EnhancedSamplingExperiments
 
 openmm_parameters = {'duration': '50ns',
                      'savefreq': '50ps',
-                     'stepsize': '1fs',
+                     'stepsize': '1.5fs',
                      'frictioncoeff': '1ps',
                      'precision': 'mixed',
                      'watermodel': 'tip3p',
@@ -20,7 +21,11 @@ openmm_parameters = {'duration': '50ns',
                      'forcefield': 'amber14',
                      'equilibrate': 'NVT',
                      'integrator': 'LangevinBAOAB',
-                     'gpu': 0,
+                     'gpu': '0',
+                     'num_water': None,
+                     'equilibration_length': '0.1ns',
+                     'ionic_strength': '0.0mol',
+                     'state_data': True,
                      }
 meta_d_parameters = {'gaussian_height': 1.2,
                      'gaussian_pace': 500,  # 500 * stepsize = 1ps
@@ -37,8 +42,7 @@ if __name__ == "__main__":
     # Alanine dipeptide
     output_dir = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/enhanced_sampling"
     starting_structures = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/one_configuration"
-    openmm_parameters[
-        'pdb'] = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/5us_NPT_alanine/top_1ps.pdb"
+    openmm_parameters['pdb'] = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/5us_NPT_alanine/top_1ps.pdb"
 
     phi = [4, 6, 8, 14]
     psi = [6, 8, 14, 16]
@@ -47,42 +51,32 @@ if __name__ == "__main__":
     feature_list = np.array([phi, psi, zeta, theta])
 
     atoms_in_alanine_dipeptide = 22
-    extended_feature_list = features.create_torsions_list(atoms_in_alanine_dipeptide,
-                                                                         size=96,
-                                                                         print_list=False,
-                                                                         append_to=list(feature_list))
-    initial_features = len(extended_feature_list)
-    extended_feature_list = np.unique(extended_feature_list, axis=0)
-    final_features = len(extended_feature_list)
-    print(f"Initial features: {initial_features}, final features: {final_features}")
+    if os.path.isfile(f"{output_dir}/100_feature_list.txt"):
+        with open(f"{output_dir}/100_feature_list.txt", 'r') as f:
+            lines = f.readlines()
 
-    # write the feature list to a file
-    with open(f"{output_dir}/100_feature_list.txt", "w") as f:
-        for feature in extended_feature_list:
-            f.write(f"{feature}")
+        data = []
+        for line in lines:
+            row = [int(x) for x in line.strip().replace('[', '').replace(']', '').split()]
+            data.append(row)
+        final_features = np.array(data)
+    else:
+        final_features = features.create_torsions_list(atoms_in_alanine_dipeptide,
+                                                       size=96,
+                                                       print_list=False,
+                                                       append_to=list(feature_list))
+        final_features = np.unique(final_features, axis=0)
 
-    # PHI-PSI Enhanced Sampling (NVT): very long run
-    # openmm_parameters['--duration'] = '250ns'
-    # CVs = ['DIH: ACE 1 C 4 0 - ALA 2 N 6 0 - ALA 2 CA 8 0 - ALA 2 C 14 0 ', 'DIH: ALA 2 N 6 0 - ALA 2 CA 8 0 - ALA 2 C 14 0 - NME 3 N 16 0 ']
-    # alanine_exp = EnhancedSamplingExperiments(
-    #     output_dir=output_dir,
-    #     unbiased_exp=alanine_system,
-    #     CVs=CVs,
-    #     starting_structures=starting_structures,
-    #     number_of_repeats=1,
-    #     openmm_parameters=openmm_parameters,
-    #     meta_d_parameters=meta_d_parameters,
-    #     features=feature_list,
-    #     subtract_feature_means=False,
-    #     feature_dimensions=2,
-    # )
-    # alanine_exp.initialise_hills_and_PLUMED()
-    # alanine_exp.run_openmm_experiments()
-    # openmm_parameters['--duration'] = '50ns'
+        # write the feature list to file
+        with open(f"{output_dir}/100_feature_list.txt", "w") as f:
+            for feature in final_features:
+                f.write(f"{feature}\n")
 
-    starting_structures = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/open_configurations"
-    #
-    # # PCA Enhanced Sampling (NVT) - 4 features
+    print(f"Total #features in final features: {len(final_features)}")
+
+    starting_structures = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/one_configuration"
+
+    # PCA Enhanced Sampling (NVT) - 4 features
     # CVs = ['PCA:0', 'PCA:1']
     # alanine_exp = EnhancedSamplingExperiments(
     #     output_dir=output_dir,
@@ -93,14 +87,14 @@ if __name__ == "__main__":
     #     openmm_parameters=openmm_parameters,
     #     meta_d_parameters=meta_d_parameters,
     #     features=feature_list,
-    #     cos_sin=False,
+    #     cos_sin=True,
     #     subtract_feature_means=False,
     #     num_cv_features=2,
     # )
     # alanine_exp.initialise_hills_and_PLUMED()
     # alanine_exp.run_openmm_experiments()
-    # #
-    # # # TICA Enhanced Sampling (NVT) - 4 features
+
+    # # TICA Enhanced Sampling (NVT) - 4 features
     # CVs = ['TICA:0', 'TICA:1']
     # alanine_exp = EnhancedSamplingExperiments(
     #     output_dir=output_dir,
@@ -118,8 +112,8 @@ if __name__ == "__main__":
     # )
     # alanine_exp.initialise_hills_and_PLUMED()
     # alanine_exp.run_openmm_experiments()
-    # #
-    # # # VAMP Enhanced Sampling (NVT) - 4 features
+
+    # VAMP Enhanced Sampling (NVT) - 4 features
     # CVs = ['VAMP:0', 'VAMP:1']
     # alanine_exp = EnhancedSamplingExperiments(
     #     output_dir=output_dir,
@@ -142,44 +136,42 @@ if __name__ == "__main__":
     # 100 feature experiments
     ###################################################
 
-    # starting_structures = "/home/dominic/PycharmProjects/CV_learning/exp/outputs/alanine_dipeptide/open_configurations"
-    #
     # PCA Enhanced Sampling (NVT) - 100 features
-    CVs = ['PCA:0', 'PCA:1']
-    alanine_exp = EnhancedSamplingExperiments(
-        output_dir=output_dir,
-        unbiased_exp=alanine_system,
-        CVs=CVs,
-        starting_structures=starting_structures,
-        number_of_repeats=1,
-        openmm_parameters=openmm_parameters,
-        meta_d_parameters=meta_d_parameters,
-        features=extended_feature_list,
-        subtract_feature_means=False,
-        cos_sin=True,
-        num_cv_features=4,
-    )
-    alanine_exp.initialise_hills_and_PLUMED()
-    alanine_exp.run_openmm_experiments()
+    # CVs = ['PCA:0', 'PCA:1']
+    # alanine_exp = EnhancedSamplingExperiments(
+    #     output_dir=output_dir,
+    #     unbiased_exp=alanine_system,
+    #     CVs=CVs,
+    #     starting_structures=starting_structures,
+    #     number_of_repeats=1,
+    #     openmm_parameters=openmm_parameters,
+    #     meta_d_parameters=meta_d_parameters,
+    #     features=final_features,
+    #     subtract_feature_means=False,
+    #     cos_sin=True,
+    #     num_cv_features=4,
+    # )
+    # alanine_exp.initialise_hills_and_PLUMED()
+    # alanine_exp.run_openmm_experiments()
 
     # TICA Enhanced Sampling (NVT) - 100 features
-    CVs = ['TICA:0', 'TICA:1']
-    alanine_exp = EnhancedSamplingExperiments(
-        output_dir=output_dir,
-        unbiased_exp=alanine_system,
-        CVs=CVs,
-        starting_structures=starting_structures,
-        number_of_repeats=1,
-        openmm_parameters=openmm_parameters,
-        meta_d_parameters=meta_d_parameters,
-        features=extended_feature_list,
-        subtract_feature_means=False,
-        cos_sin=True,
-        num_cv_features=4,
-        lagtime=1,
-    )
-    alanine_exp.initialise_hills_and_PLUMED()
-    alanine_exp.run_openmm_experiments()
+    # CVs = ['TICA:0', 'TICA:1']
+    # alanine_exp = EnhancedSamplingExperiments(
+    #     output_dir=output_dir,
+    #     unbiased_exp=alanine_system,
+    #     CVs=CVs,
+    #     starting_structures=starting_structures,
+    #     number_of_repeats=1,
+    #     openmm_parameters=openmm_parameters,
+    #     meta_d_parameters=meta_d_parameters,
+    #     features=final_features,
+    #     subtract_feature_means=False,
+    #     cos_sin=True,
+    #     num_cv_features=4,
+    #     lagtime=1,
+    # )
+    # alanine_exp.initialise_hills_and_PLUMED()
+    # alanine_exp.run_openmm_experiments()
 
     # VAMP Enhanced Sampling (NVT) - 100 features
     CVs = ['VAMP:0', 'VAMP:1']
@@ -191,7 +183,7 @@ if __name__ == "__main__":
         number_of_repeats=1,
         openmm_parameters=openmm_parameters,
         meta_d_parameters=meta_d_parameters,
-        features=extended_feature_list,
+        features=final_features,
         subtract_feature_means=False,
         cos_sin=True,
         num_cv_features=4,
